# api/openapi.yaml
openapi: 3.0.3
info:
  title: flockd local API
  description: |
    The Teraflock node daemon's loopback HTTP surface on `localhost:7777`.

    Two halves:
    - **management** (`/api/v1/*`): the REST+SSE API behind the CLI, TUI,
      embedded web dashboard, and desktop app. Bearer-token auth (the token
      lives at `<data_dir>/local_api_token`, 0600 — same-user access is the
      security model). This spec **generates the daemon's router and types**
      (oapi-codegen, `make gen`): it cannot drift from the implementation.
    - **openai** (`/v1/*`): the OpenAI-compatible inference surface. Wire
      format follows the OpenAI API; endpooints are documented here for
      completeness but their schemas are owned by that external standard.
      Keyless on loopback unless `local_api.require_auth_v1` is set.
  version: "1.0"
servers:
  - url: http://127.0.0.1:7777
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/status:
    get:
      tags: [management]
      operationId: getStatus
      summary: Node status snapshot
      responses:
        "200":
          description: Current node status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/models:
    get:
      tags: [management]
      operationId: listModels
      summary: Locally installed models
      responses:
        "200":
          description: Cache contents merged with runtime state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModelList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/models/{id}/pin:
    post:
      tags: [management]
      operationId: pinModel
      summary: Pin or unpin a model (pinned models are never LRU-evicted)
      parameters: [{ $ref: "#/components/parameters/ModelID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PinRequest" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "501": { $ref: "#/components/responses/Unsupported" }
...
  /api/v1/models/{id}:
    delete:
      tags: [management]
      operationId: deleteModel
      summary: Delete a model from disk (and unload it if loaded)
      parameters: [{ $ref: "#/components/parameters/ModelID" }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "501": { $ref: "#/components/responses/Unsupported" }
  /api/v1/earnings:
    get:
      tags: [management]
      operationId: getEarnings
      summary: Earnings (ledger-backed when enrolled, else a labelled estimate)
      responses:
        "200":
          description: Earnings snapshot.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Earnings" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/limits:
    get:
      tags: [management]
      operationId: getLimits
      summary: Live governor limits
      responses:
        "200":
          description: Active limits.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Limits" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "501": { $ref: "#/components/responses/Unsupported" }
    put:
      tags: [management]
      operationId: updateLimits
      summary: Update governor limits
      description: >
        Applies live and persists to the daemon-owned
        `<data_dir>/limits.toml` overlay; the operator's config.toml is
        never rewritten.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Limits" }
      responses:
        "200":
          description: The applied limits.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Limits" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "501": { $ref: "#/components/responses/Unsupported" }
  /api/v1/logs:
    get:
      tags: [management]
      operationId: getLogs
      summary: Recent daemon log entries (ring buffer)
      parameters:
        - name: n
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 1024 }
          description: Max entries to return (default 200). `limit` is an accepted alias.
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 1024 }
          description: Alias for `n`.
      responses:
        "200":
          description: Oldest-first log entries.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/activity:
    get:
      tags: [management]
      operationId: getActivity
      summary: Recent node activity (downloads, loads, evictions, declines)
      description: >
        The last 200 things the daemon did to its model store and runtimes,
        newest first, including everything the mesh did on the operator's
        behalf. In memory only; cleared on restart.
      responses:
        "200":
          description: Activity rows.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActivityList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/update/check:
    post:
      tags: [management]
      operationId: checkUpdate
      summary: Run a version check now
      description: >
        Fetches the version feed immediately (the daemon otherwise checks
        on startup and hourly) and returns the result.
      responses:
        "200":
          description: Check result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Update" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/Upstream" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Contents of `<data_dir>/local_api_token`.
  parameters:
    ModelID:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Model id (catalog id or cache id).
  responses:
    Ok:
      description: Success.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Ok" }
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Unknown model or resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unsupported:
      description: Not available on this node (mock runtime, or standalone).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    Status:
      type: object
      required:
        [node_id, version, standalone, enrolled, state, uptime_seconds,
         default_model, models_loaded, inflight, on_battery, temp_celsius,
         stats, memory, disk]
      properties:
        node_id:
          type: string
          description: Key fingerprint until enrolled; coordinator-assigned id after.
        version: { type: string }
        standalone: { type: boolean }
        enrolled: { type: boolean }
        cert_expires_at:
          type: string
          format: date-time
          description: Client-certificate expiry; present when enrolled.
        cert_rotation_error:
          type: string
          description: >
            Why the last due client-certificate rotation failed; absent when
            rotation is not due or succeeded. The node keeps serving on the
            current certificate and leaves the mesh when it expires unless a
            later restart rotates it or the node is re-enrolled.
        state:
          type: string
          description: >
            serving | starting | idle | no-model | yielded | paused-battery |
            paused-thermal | outside-schedule. `starting`: the default model
            is still being fetched or loaded; `idle`: nothing loaded but
            models are on disk (a placement or request loads one);
            `no-model`: nothing loaded and nothing on disk.
        uptime_seconds: { type: integer, format: int64 }
        default_model: { type: string }
        models_loaded: { type: integer }
        inflight: { type: integer }
        on_battery: { type: boolean }
        temp_celsius: { type: number, format: double }
        hardware: { $ref: "#/components/schemas/Hardware" }
        runtime_accel:
          type: string
          description: >
            Accelerator backend of the llama-server build this node runs
            (`cuda12`, `rocm`, `vulkan`, `metal`, `cpu-avx2`, `cpu`): the
            first lane of the detection preference chain
            (cuda12 > rocm > vulkan > cpu) that the pinned runtime manifest
            publishes, so it can differ from `hardware.gpus[].accel` —
            an AMD box serves on `vulkan` while no ROCm build is published.
            Absent for the mock runtime.
        stats: { $ref: "#/components/schemas/Stats" }
        memory: { $ref: "#/components/schemas/Memory" }
        disk: { $ref: "#/components/schemas/Disk" }
        update:
          $ref: "#/components/schemas/Update"
          description: Present once the first version check has completed.
...
    Update:
      type: object
      description: >
        Result of the last version check against the mesh's version feed.
        `minimum` is the oldest daemon the coordinator still serves; below
        it the node is drained until updated.
      required: [available, current, latest, checked_at]
      properties:
        available: { type: boolean }
        current: { type: string }
        latest: { type: string }
        minimum: { type: string }
        below_minimum: { type: boolean }
        url:
          type: string
          description: Release page / download for `latest`.
        checked_at: { type: string, format: date-time }
...
    ModelList:
      type: object
      required: [models]
      properties:
        models:
          type: array
          items: { $ref: "#/components/schemas/ModelRow" }
...
    Earnings:
      type: object
      description: >-
        Earnings as this node knows them. `source: ledger` — the figures are
        the operator account's ledger position, pushed by the coordinator
        over the tunnel (every node of the operator together; settled =
        vested and redeemable, pending = in escrow until the canary window
        clears). `source: estimated` — no fresh ledger snapshot (standalone,
        not enrolled, or the tunnel has been down for two push intervals):
        a simulated figure from this node's own counters, labelled in
        `note`. The `*_credits` integer fields are ledger credits at
        `credits_per_usd`; the legacy `earned_credits` / `escrow_credits`
        doubles are those divided by 1e6.
      required:
        [source, earned_microcredits, earned_credits, est_usd, est_usd_per_day,
         lifetime_tokens, escrow_credits, note]
      properties:
        source:
          type: string
          enum: [ledger, estimated]
        as_of:
          type: string
          format: date-time
          description: When the coordinator read the ledger (ledger only).
        settled_credits:
          type: integer
          format: int64
          description: Vested, redeemable balance (ledger only).
        pending_credits:
          type: integer
          format: int64
          description: In escrow, vesting after the canary window (ledger only).
        earned_today_credits:
          type: integer
          format: int64
          description: Payout credits earned since 00:00 UTC (ledger only).
        earned_7d_credits:
          type: integer
          format: int64
          description: Payout credits earned in the trailing seven days (ledger only).
        lifetime_payout_credits:
          type: integer
          format: int64
          description: Payout credits earned since the account opened (ledger only).
        credits_per_usd:
          type: integer
          format: int64
          description: The coordinator's peg; dollars are credits divided by this (ledger only).
        next_vest_at:
          type: string
          format: date-time
          description: When the oldest pending lot vests; absent with nothing pending.
        earned_microcredits: { type: integer, format: int64 }
        earned_credits: { type: number, format: double }
        est_usd: { type: number, format: double }
        est_usd_per_day: { type: number, format: double }
        lifetime_tokens:
          type: integer
          format: int64
          description: Always this node's own count, whatever the source.
        escrow_credits: { type: number, format: double }
        note: { type: string }
...
    ActivityList:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items: { $ref: "#/components/schemas/ActivityEvent" }
...
    LogList:
      type: object
      required: [logs]
      properties:
        logs:
          type: array
          items: { $ref: "#/components/schemas/LogEntry" }
    PinRequest:
      type: object
      required: [pinned]
      properties:
        pinned: { type: boolean }
...